test(runtime): isolate localized fault regressions
