+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+- 2026-08-22: Continued #2564 test-module refactoring by moving the focused
+  localized runtime-fault regressions into
+  `test_prg_engine_control_flow_localized_faults.cpp`. The existing
+  `test_prg_engine_control_flow` executable, invocation order, assertions, and
+  runtime behavior remain unchanged.
+
 - 2026-08-22: Continued #2564 test-module refactoring by moving the cohesive
   row-source type, bound-column, column-count, list-index, left-column, and
   display-value Studio-host selector regressions into
